create new users with an updated argon2 hash config and password
