+++ server/keys/symmetric.go
@@ -4,14 +4,10 @@
 package keys
 
 import (
-	"bytes"
 	"crypto/aes"
 	"crypto/cipher"
 	"crypto/rand"
-	"errors"
 	"fmt"
-
-	"golang.org/x/crypto/argon2"
 )
 
 // GenerateRandomBytes generates a slice of bytes of the given size that is
@@ -70,72 +66,4 @@ func DecryptWith(key []byte, s string) ([]byte, error) {
 		return nil, fmt.Errorf("keys: error unmarshaling cipher: %w", err)
 	}
 	return aesgcm.Open(nil, v.nonce, v.cipher, nil)
-}
-
-// DeriveKey wraps package argon2 in order to derive a symmetric key from the
-// given value (most likely a password) and the given salt.
-func DeriveKey(value, encodedSalt string) ([]byte, error) {
-	salt, saltErr := unmarshalVersionedCipher(encodedSalt)
-	if saltErr != nil {
-		return nil, fmt.Errorf("keys: error decoding salt into bytes: %w", saltErr)
-	}
-	switch salt.algoVersion {
-	case passwordAlgoArgon2:
-		key := defaultArgon2Hash([]byte(value), salt.cipher, DefaultEncryptionKeySize)
-		return key, nil
-	default:
-		return nil, fmt.Errorf("keys: received unknown algo version %d for deriving key", salt.algoVersion)
-	}
-}
-
-const (
-	passwordAlgoArgon2 = 1
-)
-
-// NewSalt creates a new salt value of the default length and wraps it in a
-// versioned cipher
-func NewSalt() (*VersionedCipher, error) {
-	b, err := GenerateRandomBytes(DefaultSaltLength)
-	if err != nil {
-		return nil, fmt.Errorf("keys: error generating random salt: %w", err)
-	}
-	return newVersionedCipher(b, passwordAlgoArgon2), nil
-}
-
-// HashString hashed the given string using argon2
-func HashString(pw string) (*VersionedCipher, error) {
-	if pw == "" {
-		return nil, errors.New("keys: cannot hash an empty string")
-	}
-	salt, saltErr := GenerateRandomBytes(DefaultSecretLength)
-	if saltErr != nil {
-		return nil, fmt.Errorf("keys: error generating random salt for password hash: %w", saltErr)
-	}
-	hash := defaultArgon2Hash([]byte(pw), salt, DefaultPasswordHashSize)
-	return newVersionedCipher(hash, passwordAlgoArgon2).addNonce(salt), nil
-}
-
-// CompareString compares a string with a stored hash
-func CompareString(password, cipher string) error {
-	if cipher == "" {
-		return errors.New("keys: cannot compare against an empty cipher")
-	}
-	v, err := unmarshalVersionedCipher(cipher)
-	if err != nil {
-		return fmt.Errorf("keys: error parsing versioned cipher: %w", err)
-	}
-	switch v.algoVersion {
-	case passwordAlgoArgon2:
-		hash := defaultArgon2Hash([]byte(password), v.nonce, DefaultPasswordHashSize)
-		if bytes.Compare(hash, v.cipher) != 0 {
-			return errors.New("keys: could not match passwords")
-		}
-		return nil
-	default:
-		return fmt.Errorf("keys: received unknown algo version %d for comparing passwords", v.algoVersion)
-	}
-}
-
-func defaultArgon2Hash(val, salt []byte, size uint32) []byte {
-	return argon2.IDKey(val, salt, 1, 64*1024, 4, size)
 }
