+++ server/keys/symmetric_test.go
@@ -47,16 +47,3 @@ func TestSymmetricEncryption(t *testing.T) {
 		})
 	}
 }
-
-func TestHashString(t *testing.T) {
-	hash, hashErr := HashString("s3cr3t")
-	if hashErr != nil {
-		t.Fatalf("Unexpected error %v", hashErr)
-	}
-	if err := CompareString("s3cr3t", hash.Marshal()); err != nil {
-		t.Errorf("Unexpected error %v", err)
-	}
-	if err := CompareString("other", hash.Marshal()); err == nil {
-		t.Errorf("Comparison unexpectedly passed for wrong password")
-	}
-}
